# api-docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "GeoCrash DE",
    "description": "Spatial Analysis of Traffic Accidents in Germany \u2014 TU Chemnitz DWT Project",
    "license": {
      "name": "dl-de/by-2-0",
      "url": "https://www.govdata.de/dl-de/by-2-0"
    },
    "version": "0.1.0"
  },
  "paths": {
    "/regions": {
      "get": {
        "tags": [
          "regions"
        ],
        "summary": "List Regions",
        "operationId": "list_regions_regions_get",
        "parameters": [
          {
            "name": "level",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "pattern": "^(state|district|municipality)$"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Level"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/regions/{ags}/indicators": {
      "get": {
        "tags": [
          "regions"
        ],
        "summary": "Get Region Indicators",
        "description": "Return all indicator time-series for a region.",
        "operationId": "get_region_indicators_regions__ags__indicators_get",
        "parameters": [
          {
            "name": "ags",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "examples": [
                "11000"
              ],
              "title": "Ags"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/regions/{ags}": {
      "get": {
        "tags": [
          "regions"
        ],
        "summary": "Get Region",
        "operationId": "get_region_regions__ags__get",
        "parameters": [
          {
            "name": "ags",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "examples": [
                "11000"
              ],
              "title": "Ags"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}